bugfix-wrong uid and gid setting in container
Previously, if we don't set the uid and the gid of the dockeruser to be
the same as the uid and gid in the host system, we cannot create or
modify any files in the container.

Now, the build_image.sh will automatically determine the uid and gid of
the current user and use it config the dockeruser we create in the
container.