remote inheritance from enable_shared_from_this
It seems that this inheritance is not needed, since I typically will not
create another new share_ptr instances within c++ from an existing env
object. This is because the env object is normally send to python side
and is handled by the python.

In python, if I just want to pass around this env object, python will
simply increases the reference count of the underlying shared_ptr.